+++ klause/src/commonTest/kotlin/com/eignex/klause/lp/cut/AggregationMirSeparatorTest.kt
@@ -130,7 +130,7 @@ class AggregationMirSeparatorTest {
 
     private fun satisfiesPb(f: PseudoBoolean, bools: BooleanArray): Boolean {
         var s = 0L
-        for (k in f.literals.indices) s += f.weights[k].toLong() * litValue(f.literals[k], bools)
+        for (k in f.literals.indices) s += f.weights[k] * litValue(f.literals[k], bools)
         return when (f.op) {
             PbOp.LE -> s <= f.bound
             PbOp.GE -> s >= f.bound
@@ -166,11 +166,11 @@ class AggregationMirSeparatorTest {
                 val lits = IntArray(k) { j -> Lit.make(vars[j], rng.nextBoolean()) }
                 when (rng.nextInt(3)) {
                     0 -> factors.add(
-                        PseudoBoolean(IntArray(k) { rng.nextInt(1, 4) }, lits, PbOp.LE, rng.nextInt(0, 2 * k)),
+                        PseudoBoolean(LongArray(k) { rng.nextInt(1, 4).toLong() }, lits, PbOp.LE, rng.nextInt(0, 2 * k).toLong()),
                     )
 
                     1 -> factors.add(
-                        PseudoBoolean(IntArray(k) { rng.nextInt(1, 4) }, lits, PbOp.GE, rng.nextInt(0, 2 * k)),
+                        PseudoBoolean(LongArray(k) { rng.nextInt(1, 4).toLong() }, lits, PbOp.GE, rng.nextInt(0, 2 * k).toLong()),
                     )
 
                     else -> {
